changed implementation of telemetry server
