+++ Cargo.toml
@@ -17,7 +17,7 @@ tokio-stream = "0.1.11"
 futures-util = "0.3.17"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.68"
-hyper = { version = "0.14", features = ["full"]}
+hyper = { version = "0.14.23", features = ["full", "http1"]}
 hyper-tls = "0.5.0"
 rand = "0.8.4"
 regex = "1.5"
